Replaced with a new version meant to run on a ESP8266 (D1 Mini) since for some reason the previous one failed to work reliable - only worked once when it actually could recognise the scanners, but worked intermittently when we moved to 3.3V powered systems.
When the correct series of cards are scanned (determined by the array correct_uids), then RX is activated for OUTPUT_TIME ms or permanently if OUTPUT_TIME is configured to be -1.

Rather than the correct series of cards having to all be scanned in one shot, each reader just has to see its correct card in the last LAST_SEEN_THRESHOLD ms. This makes the system more responsive to the correct series of cards being scanned.

Scanning a card with the id reset_uid will immediately deactivate the
output pin